Рефакторит actions
