+++ project/src/components/common/FormReview/FormReview.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {useHistory} from 'react-router-dom';
 
-import { addReview } from '../../../store/actions';
+import { addReview } from '../../../store/api-actions';
 import {APIRoute} from '../../../const';
 
 function FormReview({ onSubmit, id }) {
