+++ project/src/index.js
@@ -6,13 +6,13 @@ import {reducer} from './store/reducer';
 import thunk from 'redux-thunk';
 import {composeWithDevTools} from 'redux-devtools-extension';
 
-import {ActionCreator, checkAuth} from './store/actions';
+import {checkAuth, requireAuthorization} from './store/actions';
 import {createAPI} from './services/api';
 import {AuthorizationStatus} from './const';
 import App from './components/App/app';
 
 const api = createAPI(
-  () => store.dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH)),
+  () => store.dispatch(requireAuthorization(AuthorizationStatus.NO_AUTH)),
 );
 
 const store = createStore(
